Extract TargetTriple into its own file target_triple.rs
