+++ src/commands/toolchain.rs
@@ -4,7 +4,8 @@ use clap::Parser;
 use crate::ops::fuelup_toolchain::install::install;
 use crate::ops::fuelup_toolchain::new::new;
 use crate::ops::fuelup_toolchain::uninstall::uninstall;
-use crate::toolchain::{TargetTriple, RESERVED_TOOLCHAIN_NAMES};
+use crate::target_triple::TargetTriple;
+use crate::toolchain::RESERVED_TOOLCHAIN_NAMES;
 
 #[derive(Debug, Parser)]
 pub enum ToolchainCommand {
