+++ src/ops/fuelup_toolchain/install.rs
@@ -2,7 +2,8 @@ use crate::component;
 use crate::download::DownloadCfg;
 use crate::path::settings_file;
 use crate::settings::SettingsFile;
-use crate::toolchain::{DistToolchainName, TargetTriple, Toolchain};
+use crate::target_triple::TargetTriple;
+use crate::toolchain::{DistToolchainName, Toolchain};
 use crate::{channel::Channel, commands::toolchain::InstallCommand};
 use anyhow::Result;
 use std::fmt::Write;
